requirements.txt
psycopg2
fastapi
pydantic
dotenv
uvicorn
azure-storage-blob